Split TwoDBasis into a separate file
