+++ src/CMakeLists.txt
@@ -6,6 +6,7 @@ general/timer.cpp general/elements.cpp general/lobatto.cpp
 general/angular.cpp general/scf_helpers.cpp general/lcao.cpp
 general/gsz.cpp general/sap.cpp general/dftfuncs.cpp
 general/checkpoint.cpp atomic/basis.cpp atomic/quadrature.cpp
+atomic/TwoDBasis.cpp
 atomic/dftgrid.cpp atomic/erfc_expn.cpp sadatom/basis.cpp
 sadatom/dftgrid.cpp sadatom/solver.cpp sadatom/configurations.cpp
 general/dftfuncs.cpp diatomic/basis.cpp diatomic/quadrature.cpp
