+++ src/general/checkpoint.h
@@ -19,6 +19,7 @@
 
 #include <armadillo>
 #include "../atomic/basis.h"
+#include "../atomic/TwoDBasis.h"
 #include "../diatomic/basis.h"
 
 // Use C routines, since C++ routines don't seem to add any ease of use.
