+++ src/sadatom/solver.h
@@ -20,6 +20,7 @@
 #include <armadillo>
 #include "dftgrid.h"
 #include "../atomic/basis.h"
+#include "../atomic/TwoDBasis.h"
 #include "../atomic/dftgrid.h"
 
 namespace helfem {
